pyproject.toml
[project]
name = "parallel"
version = "0.1.0"
description = "Demonstrating parallel patterns with CrewAI"
authors = [{ name = "Rick Jang", email = "jyjang@uengine.org" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai>=0.117.1",
    "crewai-tools[mcp]>=0.43.0",
    "langchain-community>=0.3.23",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.2.14",
    "langgraph>=0.3.34",
    "mcp>=1.6.0",
    "mem0ai>=0.1.94",
    "python-dotenv>=1.1.0",
    "supabase>=2.0.0",
    "unstructured>=0.17.2",
    "psycopg2-binary>=2.9.9",  # PostgreSQL 드라이버
    "fastapi>=0.109.0",        # FastAPI
    "uvicorn>=0.27.0",         # ASGI 서버
    "pydantic>=2.6.0",         # 데이터 검증
    "firebase-admin>=6.4.0",   # Firebase
    "pytz>=2024.1",           # 시간대 처리
    "diff-match-patch>=20230430",  # 정확한 텍스트 diff 처리
]

[project.scripts]
parallel_workflows = "crewai_parallel_patterns.main:run"
run_crew = "parallel.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"